Вынес весь текс в ресурсы, добавил инициализацию ViewBinding для всех фрагментов
