+++ app/src/main/java/ru/practicum/android/diploma/ui/screens/FavouritesFragment.kt
@@ -5,14 +5,25 @@ import android.view.LayoutInflater
 import android.view.View
 import android.view.ViewGroup
 import androidx.fragment.app.Fragment
-import ru.practicum.android.diploma.R
+import ru.practicum.android.diploma.databinding.FragmentFavouritesBinding
 
 class FavouritesFragment : Fragment() {
+
+    private var _binding: FragmentFavouritesBinding? = null
+    private val binding: FragmentFavouritesBinding get() = requireNotNull(_binding)
+
     override fun onCreateView(
         inflater: LayoutInflater,
         container: ViewGroup?,
         savedInstanceState: Bundle?
-    ): View? {
-        return inflater.inflate(R.layout.fragment_favourites, container, false)
+    ): View {
+        _binding = FragmentFavouritesBinding.inflate(inflater, container, false)
+        return binding.root
     }
+
+    override fun onDestroyView() {
+        super.onDestroyView()
+        _binding = null
+    }
+
 }
