--- app/src/main/java/ru/practicum/android/diploma/ui/team/TeamFragment.kt
@@ -1,29 +1,0 @@
-package ru.practicum.android.diploma.ui.team
-
-import android.os.Bundle
-import android.view.LayoutInflater
-import android.view.View
-import android.view.ViewGroup
-import androidx.fragment.app.Fragment
-import ru.practicum.android.diploma.databinding.FragmentTeamBinding
-
-class TeamFragment : Fragment() {
-
-    private var _binding: FragmentTeamBinding? = null
-    private val binding: FragmentTeamBinding get() = requireNotNull(_binding)
-
-    override fun onCreateView(
-        inflater: LayoutInflater,
-        container: ViewGroup?,
-        savedInstanceState: Bundle?
-    ): View {
-        _binding = FragmentTeamBinding.inflate(inflater, container, false)
-        return binding.root
-    }
-
-    override fun onDestroyView() {
-        super.onDestroyView()
-        _binding = null
-    }
-
-}
