+++ app/src/main/res/layout/fragment_team.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".ui.team.TeamFragment"
+    tools:context=".ui.screens.TeamFragment"
     android:orientation="vertical">
 
         android:text="@string/team"
