+++ app/src/main/res/navigation/nav_graph.xml
@@ -29,4 +29,4 @@
         android:label="fragment_team"
         tools:layout="@layout/fragment_team" />
 
-</navigation> 
+</navigation>
