+++ app/src/main/res/values/strings.xml
@@ -5,4 +5,9 @@
     <string name="alexander_antipin">Антипин Александр</string>
     <string name="sergey_boikov">Бойков Сергей</string>
     <string name="sukhanov_temirlan">Суханов Темирлан</string>
+    <string name="details_screen">Vacancy Details Screen</string>
+    <string name="favourites_screen">Favourites Screen</string>
+    <string name="main_screen">Main Screen</string>
+    <string name="main">Главная</string>
+    <string name="favourite">Избранное</string>
 </resources>
